requirements.txt
absl-py==2.3.0
aiofiles==24.1.0
alembic==1.16.4
altair==5.5.0
altex==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.3
botocore==1.39.3
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.0.0
cmake==4.0.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
deepface==0.0.93
distlib==0.3.9
distro==1.9.0
dlib==20.0.0
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==37.4.0
fastapi==0.115.14
favicon==0.7.0
filelock==3.18.0
fire==0.7.0
firebase-admin==6.9.0
Flask==3.1.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
flatbuffers==25.2.10
fonttools==4.58.4
fpdf==1.7.2
gast==0.6.0
gcloud==0.18.3
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.2
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-generativeai==0.5.2
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
h5py==3.14.0
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jwcrypto==1.5.6
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==6.0.0
lz4==4.4.4
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
msgpack==1.1.1
mtcnn==1.0.0
namex==0.1.0
narwhals==1.45.0
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
openai==1.93.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.2.3
Pyrebase4==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==0.10.1
retina-face==0.0.17
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.33.0
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.41
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
uvicorn==0.35.0
validators==0.35.0
virtualenv==20.31.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
